moved CebesSparkJsonProtocol
